backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
reportlab==4.0.7
django-storages==1.14.2
django-rest-auth==0.9.5
django-allauth==0.57.0
django-notifications-hq==1.8.2
gunicorn==21.2.0
whitenoise==6.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-jose==3.3.0
django-guardian==2.4.0
django-simple-history==3.4.0
django-import-export==3.3.1
django-cleanup==8.0.0
django-environ==0.11.2
django-redis==5.4.0
django-cacheops==6.1.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3 